feat: add crud function in usePost
